refactor: rename `combobox.tsx` into `Combobox.tsx`
